Robert/habit/filter (#35)
Användaren kan nu filtrera sina rutiner efter prioritet.

Uppdaterade Metoder:

filterCategoryList - Använder sig nu av filterObjectsKeys - Metoden för att generiskt kunna filtrera ärenden utefter den kategori anvndaren vill ha filtrerad.
Flyttade metoder

buildCategoriesDropdown och buildSortDropdown har flyttats från todoBuilder till Övriga Builders (Builders.js) då de används av Habits och lär med största sannolikhet användas av events och är därför inte enbart förknippade med Todos.

* Updated method filterCategoryList to be more generic and fit into every page

* Added render-method to habits.js

* Moved Categorybuilder and sortbuilder into Builder.js

* Added documentation
