+++ JSON-storage/priorities.json
@@ -1,5 +1,10 @@
 [
     {
+        "id": 0,
+        "title": "Ingen vald...",
+        "color": "green"
+    },
+    {
         "id": 1,
         "title": "Låg",
         "color": "green"
