+++ html-pages/habits.html
@@ -60,8 +60,7 @@
           <div class="filter-group">
             <h3 class="filter-title">Prioritera</h3>
             <label for="kategori"></label>
-            <select class="dropdown">
-              <option value="">Välj prioritering</option>
+            <select id="priorities-dropdown" class="dropdown">
             </select>
           </div>
         </div>
