+++ pages-js/habits.js
@@ -3,16 +3,16 @@ import { PRIORITIES_KEY, loadFromJSONAsync } from "../services/jsonHandler.js";
 import { createHabit } from "../helpers/habitsHelper.js";
 import { buildHabit, buildHabitForm } from "../builders/habitBuilder.js";
 import { formBuilder } from "../builders/builder.js";
+import { buildCategoriesDropdownAsync } from "../builders/builder.js";
+import { filterCategoryList } from "../services/filterSortHandler.js";
 import { getInputValues, increaseDecreaseHandler } from "../services/inputHandler.js";
 
+let storage = getStorageAsJSON(HABITS_KEY);
 const createBtn = document.querySelector("[open-modal]");
 const closeModalBtn = document.querySelector("[close-modal]");
 const modal = document.querySelector("[modal]");
 
 const renderPage = () => {
-  let list = document.querySelector("section.card-container");
-  list.innerHTML = "";
-  let storage = getStorageAsJSON(HABITS_KEY);
   if (storage) {
     buildHabit(storage);
     // listItemHandler("article#todos", storage, ["description", "status", "time", "category", "deadline"]);
@@ -43,4 +43,14 @@ createBtn.addEventListener("click", async () => {
 
 closeModalBtn.addEventListener("click", () => {
   modal.close();
+});
+
+//Categories Dropdown
+let priorities = await loadFromJSONAsync(PRIORITIES_KEY);
+buildCategoriesDropdownAsync("#priorities-dropdown", priorities);
+//Event-handling
+let categoryDrop = document.querySelector("select#priorities-dropdown");
+categoryDrop.addEventListener("change", () => {
+  storage = filterCategoryList("#priorities-dropdown", HABITS_KEY, ["priority"]);
+  renderPage(storage);
 });
